Make sure end-of-file is not interpreted as a character
A schema that didn't end in a newline would enter an infinte loop eventually running out of memory.
When reaching end-of-file the get () and peek () methods returns undefined, and this was accidentaly
turned into the string 'undefined' which was then interpreted as a name start character. Now the
character class functions are properly guarded aginst undefined values